Refactor: change user mock file name
