+++ apps/api-server/src/modules/auth/auth.controller.spec.ts
@@ -4,7 +4,7 @@ import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
-import { MockUsersRepository } from '../../../test/mock/user.repository.mock';
+import { MockUsersRepository } from '../../../test/mock/users.mock';
 import { User } from '../../entities/users.entity';
 import { getRepositoryToken } from '@nestjs/typeorm';
