+++ apps/api-server/src/modules/auth/auth.service.spec.ts
@@ -3,7 +3,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { AuthService } from './auth.service';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
-import { MockUsersRepository } from '../../../test/mock/user.repository.mock';
+import { MockUsersRepository } from '../../../test/mock/users.mock';
 import { User } from '../../entities/users.entity';
 import { getRepositoryToken } from '@nestjs/typeorm';
