+++ apps/api-server/src/modules/users/users.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
 import { ConfigService } from '@nestjs/config';
-import { MockUsersRepository } from '../../../test/mock/user.repository.mock';
+import { MockUsersRepository } from '../../../test/mock/users.mock';
 import { User } from '../../entities/users.entity';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
